Work around problems with ports in t/cycle
The comments in this code suggest that a busy server can run into
problems and that's why the code retries up to 20 times. I've raised
that to 50 to deal with frequent failures in Travis CI. Unfortunately
additions to this value seem to have diminishing returns. Presumably
there are things that slip into window between checking if the port is
available and binding gearmand to it that don't let go at all.

We also change the Retry to info, instead of debug, as that is useful
information for a system administrator while the daemon starts up, and
should not be relegated to debug-only logs.

This closes #50